requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
langgraph>=0.0.40
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-google-genai>=0.0.5
pydantic>=2.0.0
python-dotenv>=1.0.0
